Refine the implementation of TCustomListView.Delete(), .InsertItem and .Edit after r64309 #fbb46cb85d (Issue #0038263).
This commit is also not completely Delphi compatible.
In Delphi Items can be added or delted as a respons to LVN_xxx messages, which the Lazarus implementation
does not support (it may very well be not protable?).

This commit causes code duplication (fortunately mostly one-liners)
The resons for this that the Lazarus methods DoDeletion, DoInsert and DoEndEdit are virtual and as such
may have been overriden by users, so this code still needs to work.
Users with code coming from Delphi expect that they can override Delete, InsertItem and Edit, so this must be possible as well,
and in that case out event dispatchers simply call the overridden Delphi variant.

The exception to this is Delete().
As far as we know, in Delphi Delete() is triggerd after an Items.Delete and acts as an event dispatcher,
but if you call it by itself, then it will actually delete the item (this in contrast to InsertItem, which only
calls the event handler and does nothing else).
(No Delphi code was studied to obtain this information, it was all done using black box testing (by LacaK and BB),
so this conclusion may be wrong.)
I decided to implement Delete in such a way that it checks wether it was called after the fact (after Items.Delete),
and if that is not the case, it will delete the item (triggering a re-entry of the method).
Not very nice, but it is in any case better then an empy implementation like it was before r64309 #fbb46cb85d.

It was discussed also to rename and deprecate our event dispatchers in favour of the Delpi ones, but I decided agains that
for the following reasons:
- naming conventions: DoXXX dispatches OnXXX
- signatures differ between DoEndEdit and Edit

git-svn-id: trunk@64318 -