Packager: Do not allow to translate lisPkgMangNewPackage ('NewPackage') string anymore. Unbreaks "Package -> New Package" and "Package -> New Component" for most translations into languages with non-latin alphabets, as new package name should be a valid Pascal identifier. Regenerated translations.
Before this change some translations of lisPkgMangNewPackage were discarded (in "Package -> New Package") or caused crashes (in "Package -> New Component") and other translations could be changed in "unexpected" ways (basically all symbols except 'a'..'z','A'..'Z','_','0'..'9' removed) in order to get a valid Pascal identifier. This is really not desired.

git-svn-id: trunk@65047 -