LazUtils: Rewritten comparison of formatting arguments of translations (in PO files). Now formatting arguments are compared in their entirety (including index, width and precision).
The reasons are:

0. Working with arguments is very prone to typos.
1. Removing index of arguments in translation can lead to exceptions.
2. Adding index to arguments in translation can lead to incorrect argument selection if e. g. '*' usage in arguments and relation of indexed and non-indexed arguments are not properly considered. Validating all this is a big complicated mess which should better be avoided.
3. Allowing to change width/precision can lead to incorrect argument rendering.

This means that swapping indexes is now effectively forbidden. This is not expected to create any unavoidable problems for translators, though. In any case such issues can better be solved by changing resource strings instead on a case by case basis.

This change allowed to spot many missed argument mismatches in different languages.