CI: explicitly clear Qt bindings caches in a separate jobs if any files in their directories are changed
Cache actually contains all files from bindings directory (both source
and object ones) and is extracted after checkout, so it overwrites
checked out sources, and thus prevents correct cache update.

Note that sources should be kept in cache, as removing them will cause
bindings to be always recomiled (both in bindings and IDE compile jobs).
It takes long time and thus undesirable.

So now a separate cache clearing job is automatically run if (and only
if) any file in bindings is changed.

Such two-job architecture allows to automatically update cache and at
the same time populate it in case of its absence (e. g. in new merge
requests).

Note that simply doing full cache rebuild if any file in bindings is
changed is not enough as this job will not be triggered e. g. for a new
merge request (if it does not modify bindings) and will cause IDE
compile jobs for it to fail. GitLab CI currently does not support
conditions for triggering a job if a cache is missing.