Issue #39988: Forbid duplicate paths in directories search list.
As the directories search list is not sorted, FPC implementation of
`TStringList` ignores the `dupIgnore` option and thus leads to having
the same entry multiple times in the list.

We forbid this manually in order to avoid searching multiple times for
the same file in the very same directory.

We also add more default search paths (Debian based).