Generate pointer types for all declared types.
If the type is not used within the unit it was declared in, then pointer
types will not generated. This causes an error when a pointer type is
used by a unit using it.

An example is a compilation error on `Pango1` that uses `Glib2` because
the type `Pgconstpointer` was not defined.

The new code generates a lost of pointer types, but at least it allows
get a code that compiles fine.