LazUtils: Remove special macOS handling from LazGetLanguageIDs procedure. Handle macOS just as any Unix.
For the record, the logic of this code was the following:
1. If the App Bundle does not do any Locale configuration, 'English' is
returned (i. e. not an ISO code).
2. If the App Bundle has configured the Locale configuration that
matches the user's system settings, it will return locale name from its
list, e. g. 'zh_CN'.
3. If the App Bundle cannot match the Locale of the user system, the
first item set in the App Bundle is returned (`en` in case of Lazarus).

So this code provided no tangible benefits and introduced dependency on
language list in Lazarus bundle. This list was outdated and thus
prevented autodetection of Czech, Hungarian, Brazilian Portuguese,
Ukrainian languages on macOS.

In case there will a need to reintroduce special macOS handling arise,
NSLocale.object() should be called with `countryCode` and `languageCode`
keys.