Cocoa: Fix GetKeyState (eg. Command/Option/Shift) that the principle is similar to 53cd3db
completely eliminate keystroke issue caused by improper reading of modifier keys.
it's because in TCocoaWidgetSet.GetKeyState(), the modifier key should not be obtained in real time, but should be obtained from the current event.
for example, when pressing Command+R in the IDE, it was originally a shortcut key for Run, but sometimes the character R is entered inadvertently.