LCL-GTK2: Fix double free in listview pixbuf due to wrong refcounting
The 'pixbuf' field in GtkCellRendererPixbuf is supposed to be private to
Gtk which handles refcounting itself, but LCL modifies it directly which
can cause refcounting mismatches, especially during the destruction of a
listview with icons (images are unref'd by both the listview destructor
and gtk itself).  Since it is exposed as a property it can be modified
using the g_object_set_property and g_value_xxx APIs which handle any
necessary refcounting, so this patch replaces the direct modification of
the pixbuf field with these APIs.  This fixes the crashes due to
refcounting mismatch.


(cherry picked from commit 6a5e16239a2ddd241af766b4a76217f22d7f5838)