LCL-GTK1: Fix menu item drawing for menus without icons
When a menu does not have any icons set, the existing code drew
checkboxes over the menu icons.  This patch fixes that.

The previous code relied on a function that apparently became noop in
Gtk 1.2.9 but the underlying code and private field it modified were not
changed, so this fix now changes the private field directly (this change
was made in 2001 so the chances of this breaking are a bit low :-P).

It also removes the unnecessary left side padding when no menu has any
icon or is checkable so that the menu appearance matches other
(non-Lazarus made) Gtk 1.2 programs.