LCL-GTK1: Fix various startup warnings and Gtk1 assertions
This fixes a bunch of warnings when Gtk1 applications start.  The fixes
are on keyboard initialization (make a previously statically sized array
dynamic as the old value wasn't long enough and disable an unnecessary
warning about filling the VK table as the user can't do anything about
it - nor we unless the whole thing is redesigned), module loading (this
is a side effect of an environment variable collision between Gtk1, Gtk2
and Gtk3 - all of these use the GTK_MODULES variable to load some
modules but since as of 2023 no distribution aside from Slackware comes
with Gtk1, all of these warnings are bogus, so this patch temporarily
cleans the environment variable before initializing Gtk and restores it
later so that child processes can still access it) and passing NULL
styles to gtk_style_copy (the previous code assumed the style retrieval
functions always return a valid object, which is not the case).