GTK3: Enable showing hot key in page control widgets caption.
We store the raw caption in the `Text` field and convert ampersands
(`&`) to underscores (`_`) before using `set_markup_with_mnemonic` to
update the label text while honoring the hot key definition.

NB: The underscore will be drawn only if user hits `alt` key which is
a Gnome/Gtk3 feature.

This fix is related to #40713.