Cocoa: fix #40081 #40853 by completely reimplementing TManualScrollView with Legacy/Overlay Style Scroller, Merge branch 'cocoa/scroll'
TCocoaManualScrollView is based on NSView without native Scroller.
it doesn't require the document to have the full size it claims, so it is usually suitable for components of unlimited size.
it corresponds to components such as TreeView, Grid, SynEdit.
Legacy and Overlay style Scroller were implemented, they have visual effects highly similar to native Scrollers.

on the contrary, TCocoaScrollView is based on NSScrollView with native Scroller.
it requires the document to have the full size it claims, so it is usually suitable for components of limited size.
it corresponds to components such as ListBox, ListView, ScrollBox, Form.