Cocoa: fix an issue related to TListBox.ScrollWidth
if column.ResizingMask is set when column.Width < TableView.Width, it is not only meaningless, but also causes problems in the vertical direction of the TableView.

we can scroll the TableView until the last row appears at the top.

it is not clear whether this is a feature of Cocoa or a bug.